move recording code to recording/mod.rs
